pyproject.toml
[tool.poetry]
name = "sentry-proxy"
version = "0.1.0"
description = ""
authors = []
maintainers = []
packages = [{ include = "sentry_proxy" }]

[tool.poetry.dependencies]
python = "^3.11"
sentry-sdk = "^1.13.0"
proxy-py = "^2.4.3"
redis = "^4.4.2"

[tool.poetry.group.dev.dependencies]
pytest = "^7.2.1"
pytest-freezegun = "^0.4.2"
mypy = "^1.0.0"
black = "^23.1.0"
isort = "^5.12.0"
pytest-cov = "^4.0.0"
types-redis = "^4.5.1.0"


[tool.pytest.ini_options]
minversion = "6.0"
norecursedirs = ["*.egg", ".eggs", "dist", "build", "docs", ".tox", ".git.", "__pycache__"]
addopts = "--strict --cov=sentry_proxy --cov-report=term:skip-covered --cov-report=html --cov-fail-under=90"
markers = "integration"


[tool.isort]
# See https://github.com/timothycrosley/isort#multi-line-output-modes
profile = "black"

[tool.mypy]
# The mypy configurations: https://mypy.readthedocs.io/en/latest/config_file.html
python_version = 3.11

warn_redundant_casts = true
warn_unused_ignores = true
disallow_untyped_calls = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
disallow_untyped_decorators = true
disallow_any_generics = true
check_untyped_defs = true
no_implicit_optional = true
ignore_errors = false
ignore_missing_imports = false
strict_optional = true
warn_unused_configs = true
disallow_any_explicit = false

[[tool.mypy.overrides]]
module = 'tests'
ignore_errors = true

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
